Add top level protocol packet description.
All other packets sit inside this packet.
Yes, there will be redundancy between this part of the spec and the
detailed sections on each packet format, as both will define the packet
kind number. I think this is acceptable. Maybe later we'll decide that
it's not, but for now the table seems useful for implementations
(perhaps we can generate it later).